+++ src/contents/ui/config/About.qml
@@ -1,0 +1,31 @@
+/**
+ * Crypto Tracker widget for KDE
+ *
+ * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
+ * @copyright 2021 Marcin Orlowski
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT
+ * @link      https://github.com/MarcinOrlowski/crypto-tracker-plasmoid
+ */
+
+import QtQuick 2.0
+import QtQuick.Controls 1.0
+import QtQuick.Layouts 1.1
+import org.kde.kirigami 2.5 as Kirigami
+import org.kde.kquickcontrols 2.0 as KQControls
+import org.kde.plasma.components 3.0 as PlasmaComponents
+import org.kde.plasma.core 2.0 as PlasmaCore
+import "../../js/crypto.js" as Crypto
+import ".."
+
+
+Kirigami.FormLayout {
+	Layout.fillWidth: true
+	id: controlRoot
+
+	AppInfo { }
+
+	Item {
+		Layout.fillWidth: true
+		Layout.fillHeight: true
+	}
+}